dock - move BarkNavRoute into bark-nav-route
