+++ src/app/user/(logged-in)/layout.tsx
@@ -1,10 +1,8 @@
 "use server";
 
 import { BarkDockLayout } from "@/components/bark/bark-dock";
-import {
-  BarkSidebarLayout,
-  BarkNavRoute,
-} from "@/components/bark/bark-sidebar";
+import { BarkSidebarLayout } from "@/components/bark/bark-sidebar";
+import { BarkNavRoute } from "@/components/bark/bark-nav-route";
 import { getAuthenticatedUserActor } from "@/lib/auth";
 import { IMG_PATH } from "@/lib/image-path";
 import { RoutePath } from "@/lib/route-path";
