+++ src/components/bark/bark-nav-route.ts
@@ -1,0 +1,6 @@
+export type BarkNavRoute = {
+  label: string;
+  href: string;
+  iconSrc?: string;
+  iconLightSrc?: string;
+};
